requirements.txt
# UI and Orchestration
streamlit==1.32.0
requests==2.31.0
Pillow==10.2.0

# Computer Vision & Vectorization
# (Note: Use CPU versions for the local hand-off)
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.5.1
numpy==1.26.4
